MA-6: extract logic of fetching contact to separate class
