+++ app/src/main/java/com/example/messengerAndroid/ui/myContacts/ContactFetcher.kt
@@ -1,0 +1,34 @@
+package com.example.messengerAndroid.ui.myContacts
+
+import android.content.Context
+import android.provider.ContactsContract
+import com.example.messengerAndroid.data.contactsRepository.contactModel.User
+
+class ContactFetcher {
+
+    fun fetchContacts(context: Context): MutableList<User> {
+        val contacts = mutableListOf<User>()
+        val cursor = context.contentResolver.query(
+            ContactsContract.CommonDataKinds.Phone.CONTENT_URI,
+            null,
+            null,
+            null,
+            null
+        )
+        cursor?.use {
+            while (it.moveToNext()) {
+                val id =
+                    it.getLong(it.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.CONTACT_ID))
+                val name =
+                    it.getString(it.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.DISPLAY_NAME))
+                val photo: String? =
+                    it.getString(it.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.PHOTO_THUMBNAIL_URI))
+                val job: String? =
+                    it.getString(it.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Organization.COMPANY))
+                val contact = User(id, photo?: "", name, job?: "")
+                contacts.add(contact)
+            }
+        }
+        return contacts
+    }
+}
